src/main/java/uk/gov/hmcts/reform/adoption/document/DocumentUtil.java
package uk.gov.hmcts.reform.adoption.document;

import uk.gov.hmcts.ccd.sdk.type.Document;
import uk.gov.hmcts.reform.adoption.document.model.AdoptionDocument;


public final class DocumentUtil {

    private DocumentUtil() {
    }

    public static Document documentFrom(final DocumentInfo documentInfo) {
        return new Document(
            documentInfo.getUrl(),
            documentInfo.getFilename(),
            documentInfo.getBinaryUrl());
    }

    public static AdoptionDocument adoptionDocumentFrom(final DocumentInfo documentInfo, final DocumentTypes documentType) {
        return AdoptionDocument
            .builder()
            .documentLink(documentFrom(documentInfo))
            .documentFileName(documentInfo.getFilename())
            .documentType(documentType)
            .build();
    }
}
